Fix G4HepEmMin/Max, use preprocessor define
instead of the present solution with decltype and templates.
The problem with the decltype solution is that std::min/max
returns a reference to the minimum/maximum, so the return value
of G4HepEmMin/Max is a reference to the respective argument of
G4HepEmMin/Max. This argument goes out of scope, however, when the
function returns.